new story book standard implimented
